test(regex): move #4884 surrogate test to grammar.rs (file-size gate)
regex.rs hit the 2000-line lint gate (2007). Relocate the new
surrogate-property test into grammar.rs alongside the js_regex_to_rust
logic it exercises, validating compilation via regex::Regex::new directly
instead of js_regexp_new. regex.rs back to 1967 lines.
